backend/services/intrados_detector.py
"""Intrados line (rib skeleton) detection service."""

import asyncio
from pathlib import Path
from typing import Dict, Any, List, Tuple
import numpy as np

try:
    from PIL import Image
    HAS_PIL = True
except ImportError:
    HAS_PIL = False

try:
    from skimage import morphology, filters
    from skimage.feature import canny
    HAS_SKIMAGE = True
except ImportError:
    HAS_SKIMAGE = False


class IntradosDetector:
    """Detect intrados lines (rib skeletons) in projection images."""
    
    def __init__(self):
        self.projections_dir = Path("./data/projections")
    
    async def detect(self, projection_id: str) -> List[Dict[str, Any]]:
        """Detect intrados lines in a projection image."""
        
        loop = asyncio.get_event_loop()
        result = await loop.run_in_executor(None, self._detect, projection_id)
        return result
    
    def _detect(self, projection_id: str) -> List[Dict[str, Any]]:
        """Internal detection (runs in thread pool)."""
        
        image_path = self.projections_dir / f"{projection_id}.png"
        
        if not image_path.exists() or not HAS_PIL:
            return self._generate_demo_lines()
        
        if not HAS_SKIMAGE:
            return self._generate_demo_lines()
        
        # Load image
        image = np.array(Image.open(image_path).convert("L"))
        
        # Edge detection
        edges = canny(image, sigma=2)
        
        # Skeletonization
        skeleton = morphology.skeletonize(edges)
        
        # Extract lines from skeleton
        lines = self._extract_lines_from_skeleton(skeleton)
        
        return lines
    
    def _extract_lines_from_skeleton(
        self,
        skeleton: np.ndarray,
    ) -> List[Dict[str, Any]]:
        """Extract line segments from a skeleton image."""
        
        lines = []
        
        # Find connected components in the skeleton
        try:
            from skimage.measure import label, regionprops
            
            labeled = label(skeleton)
            regions = regionprops(labeled)
            
            for region in regions:
                if region.area < 50:  # Skip small segments
                    continue
                
                # Get coordinates of the region
                coords = region.coords
                
                # Simplify to fewer points (subsample)
                step = max(1, len(coords) // 50)
                simplified_coords = coords[::step].tolist()
                
                lines.append({
                    "points": [(float(p[1]), float(p[0])) for p in simplified_coords],
                })
        except Exception:
            return self._generate_demo_lines()
        
        return lines if lines else self._generate_demo_lines()
    
    def _generate_demo_lines(self) -> List[Dict[str, Any]]:
        """Generate demo intrados lines for testing."""
        
        size = 512
        center = size // 2
        lines = []
        
        # Create rib-like lines
        for angle in [0, 45, 90, 135]:
            rad = np.radians(angle)
            points = []
            
            for t in np.linspace(-200, 200, 50):
                x = center + t * np.cos(rad)
                y = center + t * np.sin(rad)
                
                # Add some curvature
                curve = np.sin(t / 100) * 5
                x += curve * np.sin(rad)
                y -= curve * np.cos(rad)
                
                points.append((float(x), float(y)))
            
            lines.append({"points": points})
        
        return lines
    
    async def refine_line(
        self,
        line_id: str,
        adjustments: List[Tuple[int, float, float]],
    ) -> Dict[str, Any]:
        """Refine a detected line with manual adjustments."""
        # Implementation would adjust line points
        return {"success": True}

